+++ djangocms_frontend/apps.py
@@ -8,8 +8,10 @@ class DjangocmsFrontendConfig(apps.AppConfig):
 
     def ready(self):
         from . import plugin_tag
+        from . import component_pool
 
         plugin_tag.setup()
+        component_pool.setup()
         checks.register(check_settings)
         checks.register(check_installed_apps)
