+++ djangocms_frontend/templatetags/cms_component.py
